app/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
selenium
pdfkit
beautifulsoup4
webdriver-manager
pandas
openpyxl
xlrd
python-multipart
# Auth & Email
fastapi-mail
python-jose[cryptography]
bcrypt
pydantic>=2.0
python-dotenv
